Added snapshot testing / Moved SaveBuuton.jsx to UI/Button component
