+++ src/App.jsx
@@ -8,7 +8,7 @@ import WyChooseUs from './components/whyChooseUs/WhyChooseUs';
 import CompanyShowcaseComponent from './components/companyShowcaseComponent/CompanyShowcaseComponent';
 import { showCaseData } from './data/showCaseData';
 import './App.css';
-import SaveButton from './components/SaveButton';
+import SaveButton from './components/UI/Button/SaveButton';
 import Categories from './components/Categories/Categories';
 import Education from './components/EducationAndExperience/Education';
 import Experience from './components/EducationAndExperience/Experience';
